price

+++ migrations/m171125_231156_create_price_table.php
@@ -15,7 +15,7 @@ class m171125_231156_create_price_table extends Migration
         $this->createTable('price', [
             'id' => $this->primaryKey(),
             'name' => $this->string(64)->notNull(),
-            'price' => $this->integer()->notNull(),
+            'price' => $this->float()->notNull(),
             'service_id' => $this->integer()->notNull(),
             'index' => $this->integer()->notNull()->defaultValue(100),
         ]);
